unfixed

+++ commands/admin/serverlist.js
@@ -2,13 +2,17 @@ const Discord = require("discord.js");
 module.exports = {
   name: "serverlist",
   usage: ``,
-  category: "admin",
+  category: "owner",
   aliases: ["slt"],
   description: "Displays the list of Servers!",
   args: false,
   cooldown: 2,
-  permission: "ADMINISTRATOR",
   run: async (client, message, args) => {
+    const owner = ["740947753135243354", "767726828311543820"];
+
+    if (owner.includes(message.author.id) === false) {
+      return;
+    }
     //code
     message.delete();
     const bot = client;
